correct 64bit int define

+++ code/main.cpp
@@ -8,8 +8,8 @@
 typedef unsigned char c8;
 typedef unsigned int u32;
 typedef signed	 int i32;
-typedef unsigned long int u64;
-typedef signed	 long int i64;
+typedef unsigned long long u64;
+typedef signed	 long long i64;
 
 struct String {
 	c8 	*base;
